chore(charts): more stringent stacked bar chart test
Discovered that the stacked bar chart tests weren't actually failing when the stacked bar chart was deliberately sabotaged – it was testing that all marks had zero-or-greater values, but wasn't catching the case where something in the renderer fails and all stacked bar chart marks end up coming out with a height of zero – an empty chart that shouldn't be empty.

We don't want to remove the zero-height segments and then test for positive values for two reasons:

First, it's possible, theoretically, for a stacked bar chart to have all zero values if there's no data to display.

Second, we want to render empty values for accessibility reasons. A zero value is still information which is worth presenting to users of assistive technologies; just because a mark isn't visible doesn't mean it isn't important, that's vision-centric thinking.

Instead, let's add an assertion to the test which checks that at least one of the marks has a nonzero height; given what we know about the specifications we use for the tests, that's the expected behavior.