chore(charts): unique values for discrete scales
Duplicate values for nominal and ordinal scales should always be deduplicated, such that

['a', 'b', 'c', 'd', 'a']

becomes

['a', 'b', 'c', 'd']

This is already the case for nominal and ordinal scales as used along the y axis, as you can see below, but the same was never done for the x axis. This is likely just an oversight that developed from developing these axes separately – the x axis version was added for categorical bar charts, and the y axis version was added for dot plots, and the syntax was subtly different even though it didn't need to be. Resolving the discrepancy better positions us to merge these into a single implementation which can be used along both axes.