chore(charts): early return for color test
Given how simple it is, it does still seem to make sense to return the color scale based on examining the channel even though the general pattern we're striving for here is to prefer examining the type. In the interest of clarity, however, we can at least present that as an early return instead of a fallback case.