chore(charts): toggle position offsets
Bars need to be offset by the length of the bar when using the y encoding because in the SVG coordinate system, the y coordinates start at zero at the top and increase as you go down, somewhat counterintuitively. There's no need to do this for the x encoding, which starts at the left and increases as you go right, in the intuitive order.