fix(tooling): substitute isPresent()
Previously this used a helper function from Ember which checked for all sorts of empty-ish states like empty arrays and empty strings, but we don't actually care about that. All we are using it for is checking for nulls, so we can substitute our own tiny helper. This should be removed at some point too.