chore(tooling): ignore focus method type
TypeScript complains about this method, and in the strictest sense it is right to do so, but in practice this works and we want to use it.

An explanation taken from https://allyjs.io/tutorials/focusing-in-svg.html#focusing-svg-elements:

"In Blink and WebKit based browsers focusing SVG elements is not different from focusing an HTML element, the .focus() method is available on HTMLElement and SVGElement. Since Firefox 51 the .focus() method is also available. In Internet Explorer and Microsoft Edge this is not the case. Internet Explorer and Microsoft Edge up to version 12 allow misappropriating the .focus() method by applying the method from HTMLElement's prototype to an SVGElement: HTMLElement.prototoype.focus.apply(svgElement). Microsoft Edge 13 joined Firefox (up to version 50) in rejecting such attempts by throwing an error"