fix(tooling): add matchAll shim
Previously this used a custom equivalent of the new String.prototype.matchAll regex method which still worked even when users tried to run the code in an old and unsupported browser. Eventually this can probably be deleted once browser support is sufficiently widespread, but so long as it's still in place, relying on a public package is probably easier.