chore(tests): test multiple modules with code splitting
We've been caught between a rock and a hard place with the tests: esbuild won't concatenate for us as part of the build, and if we build each test file separately they end up with separate references to shared libraries like D3 and QUnit which then conflict at runtime. Unexpectedly, code splitting turns out to be a handy solution to this, because esbuild can detect that those libraries are shared, and instead moves them into the chunks.

This is an unorthodox way to handle test dependencies, but at least for now it seems to work.