chore(tests): look for event instead of event target in tooltip test
When simulating the mouse event on a node in memory, the event target is null instead of with the target attached as with the previous rendering method. That doesn't really matter, because the assertion was incorrect – it only needs to check that the event was included, not that the target was assigned to the event. In this sense checking the event.bubbles property is actually a more accurate proxy anyway.