chore(data): rename stackedBarData function to stackData
The stackedBarData() function was originally so named because we only needed it for stacked bar charts, and then running it for non-categorical bar charts seemed like a reasonable extension of that. Now we need to reuse the stack processing in order to render area charts, which makes the existing name inappropriate.