chore(core): default argument for scale dimensions
Scales logically require dimensions when you're computing geometry, but occasionally it's useful to sidestep that requirement, in particular when you only need to retrieve the color scale. In these cases, we call the parseScales() function with a useless dimensions object which sets both values to zero. Since the memoization helper function caches objects by reference instead of by value as with a naive approach based on JSON.stringify() might, that means it's actually not memoizing in these cases, and recomputing the scales every time it sees a new zero object. Making the zero object a default argument fixes this behavior and decreases the number of calls to the parseScales() function necessary to render a chart.