fix(core): fix memoization by reference for function arguments
Initially the memoization leaned on JavaScript's tendency to coerce objects to the idiosyncratic string value "[object Object]" with the .toString() method. This was sort of sloppy to begin with, and can be improved by more carefully setting descriptive strings for the memoization key.

Cleaning that up also reveals a problem: function arguments weren't being handled properly during memoization, and were being coerced to strings by that same .toString() method instead of being properly cached as references, which was the intention. This didn't matter in a practical sense because there aren't many higher-order functions being memoized, but it wasn't good practice because the functions would lose information from the surrounding scopes upon being memoized, which could have led to very confusing bugs if it had been allowed to continue.