chore(core): specify non-quantitative Cartesian encoding
In some cases it's enough to specify that you want the covariate, but when the covariate resolves to the color encoding it starts to become nonsensical because it doesn't makes sense to compute geometry based on color. To compute bar width in these cases, you don't want the covariate, you want the non-quantitative Cartesian encoding.