chore(core): select layer for margin rendering
When layers are only used for lightweight annotations like text and rules, it's easy to figure out which layers are the graphics and which are the annotations. This is less clear when you're superimposing multiple graphics on top of one another. In these cases, pick a layer that seems to have all the necessary encodings and then use that to compute margins.

Even if the layer that is chosen isn't necessarily the "primary" one with the "most" graphical content, the difference should be resolved internally by the unionDomains() function when the layerSpecification is being generated.