chore(core): separate radial heuristic
Separate radial from cartesian in the heuristics instead of bundling them together in a single heuristic for two-dimensional layouts. This makes the heuristics shorter and more targeted.

The reason to prioritize radial in front of cartesian is subtle: it's possible to have a radial chart with annotation text that specifies its x and y position, in which case the primary layer is the radial one, but it's effectively impossible to have the inverse, a cartesian chart with a radial annotation.