fix(core): preserve single series properties based on encoding
This once again adds a feature that was added in pull request #54 and then reverted in pull request #58 due to some unexpected compatibility hassles.

If the stack layout generator only produces a single series, as in the case of a single-color categorical bar chart, the key property is intentionally remapped to an underscore in the interest of, er, "clarity." This is because if there's no data mapping to define multiple series, the accessor for that property quietly fails and retrieves an undefined value, which is then cast to the string value "undefined" by the layout generator, and then that string is used as the key for the series. It's hideous! So when that condition is detected, it is explicitly overwritten.

But there are also times when the stack layouts produce a single series despite having an encoding for mapping series, such as when no values are present for the other values. Imagine a mapping for two values – "yes" and "no" – but all the data points provided to the stack layout generator are "yes." The stack layout generator, not knowing that the other "no" values can even exist, would detect that this is a single-series case and overwrite the "yes" key with the underscore. This is problematic as well, albeit in more subtle ways.

In order to more correctly handle both states, the remapping of undefined to underscore needs to also check that there isn't a valid mapping for multiple series, regardless of whether the input data provides data points to generate many series or only one.