chore(core): temporal bar test
Adds a new feature test for detecting temporal bar charts.

This must be distinct from the existing temporal and bar tests because of the recursive testing across layers. That is, a top level layered specification would return true for both if it contains temporal encoding and bar marks in any combination of layers, but what this test needs to detect are cases where both features are present in the same layer.