fix(core): temporal bar dimensions helper
Fiddly adjustments to range for the purpose of temporal bars must now be performed in two places: for the temporal bars themselves, and then again for any layer overlaid on top of those temporal bars, because even though they don't require the bar offset computation themselves, you want them to spatially align with the bar layer. This is easier to follow if you put it in a centralized function.