fix(tooling): mock image APIs
JSDOM doesn't supply working methods for SVG and canvas nodes, which consequently have to be mocked. The mtocks are fairly simple, and consequently any tests that rely on these methods can't be entirely trusted when they're running via the command line. This is useful for automation purposes, but in order to really be certain, you need to boot the UI for the tests so the browser can supply working versions of these methods.