fix(core): extension helper
It's likely that future features may need to make use of the Vega Lite usermeta property as a place to store instructions or configuration. In order to structure these and keep things under control, a new extension function will be used to access those, centralizing traversal of the JSON object in much the same way as the feature helper already does for standard pieces.