fix(tooling): use faster specification objects in tests
Transfer of arbitrary metadata from raw data to aggregated mark data is currently a performance bottleneck in the data.js module, since it compares each encoding field with quadratic performance, including for charts with multidimensional encoding. This could be fixed. In the meantime, a quick performance improvement is to switch to a different chart type when it's trivially possible to do so. Many tests were written against the stacked bar chart as a historical accident, but the concept they're testing holds true with other chart forms that don't depend on the slower data processing stage. This simple change alone speeds up the tests by about 15-20%.