chore(core): remove error from extension lookup
It's tempting to want to design this to be an extremely disciplined API which can emit a useful descriptive error. Unfortunately in practice this function is likely to be used as both a check and a retrieval - *if* the specification has extended usermeta, then return the extension and use it. This leaves us with a choice: either keep the error and add a separate function for hasExtension() or similar, or else remove the error so the main function can be used in conditional logic without throwing when the condition is true. The latter seems more elegant, at least for now.