fix(core): synthetic scale helper for tracking domain and range
Scales in d3 are mostly run identically to plain functions, but they also have additional methods attached which carry the domain and range, thus making them available everywhere the scale is. When using a plain function as a scale, as with the identity function, it's sometimes helpful to attach those methods as well. Doing so with a helper function makes it more predictable.