chore(core): pass specification as an argument to the tooltip event dispatcher
If the specification is not provided as a dispatch argument, it may be for the wrong layer when the dispatcher is fired. There may be a cleaner way to do this that depends on the layer scope.

The argument order is a bit unfortunate. On one hand, most interaction methods only use the event, which suggests the specification should be later in the argument list. On the other hand, the usual pattern elsewhere in the codebase is for the specification to be first. There's no clear winner. This can be sorted out later.

This also requires an additional guard because sometimes the dispatcher fires without a valid specification. Strictly speaking that shouldn't be necessary, but reality is tricky sometimes.