chore(core): use static field name with circular chart layout generator
Circular charts were using dynamic lookup based on the encoding fields with the results of the layout generator. This is not necessary – the results from the layout generator always use "key" and "value" as the property names. Ignoring them results in stable behavior as property names in the input fields vary, whereas trying to use dynamic key names resulted in curious behavior where the chart would only correctly when the property used for the quantitative value as actually called "value" so the dynamic lookup happened to match the static layout object.