fix(core): use static accessors for circular chart layouts
Using dynamic field lookup against the output of the layout generator was confounding computation with nested fields, because it was trying to use the nesting in cases where the desired value was already present on the value property. Dynamic fields are not necessary when there's a layout generator between the data and the DOM rendering.