chore(core): shared default styles object
Using default argument syntax results in the creation of a new empty object each time the function is run, which impedes memoization. Using an object from shared scope as the default means the memoization can reliably resolve. The syntax is less aesthetically pleasing, but this optimization is highly beneficial because text manipulation is among the most expensive parts of rendering, since it requires interacting with a live <canvas> node.