chore(core): calculate extents
Calculate high and low values per quantitative field for comparison. For performance reasons, this should be done once in a precursor stage before rendering the descriptions.

This means interacting with an intermediate stage of the data – after it has been aggregated by the data handling functions, but before it has been processed by the layout generator functions. The result is an awkward approximation of the accessor paradigm – functions which are accessor-like, but not quite the same as those in accessors.js. This is unfortunate, and might be worth cleaning up later.