chore(core): reusable default scale dimensions object
The dimensions argument to the parseScales() function is optional. It's mostly useful to omit this in cases where you're working with color and don't need to know anything about positioning. However, the use of default argument syntax creates a new object reference on every invocation that is consequently not memoizable. Setting this object to a reusable default is more efficient.