chore(core): use discrete helper for alternation function switch
The tick alternation function can only run when the axis is continuous, because it doesn't make sense to omit ticks entirely with a nominal or ordinal encoding in which the values cannot be interpolated. This check previously only disqualified nominal encodings, but it should also include ordinal encodings, and should do so using the isDiscrete() helper function.