fix(core): stricter isDiscrete helper
Until now this has been defined as the opposite of isContinuous(), but that's a shortcut, and strictly speaking it's not always true. This can turn up with an edge case when the encoding is undefined. Since this is so deep in the core logic, it's better to have a stricter check.