chore(core): calculate tick positions once
There's probably no need to recalculate the tick position transforms for each tick. The only thing that would even theoretically change here is the height of the text node, and that should be consistent for different pieces of text, regardless of whether they have e.g. characters with descenders like "j" and "y". Moving this out of the iteration scope and computing it once should be faster than doing it in a callback.