chore(core): select all x axis ticks for transform attribute
This was a mistake. Selecting only one tick with .select() instead of selecting all ticks with .selectAll() manifests as only the first tick in the x axis getting the transform. (This was probably what was intended for the distinction between .nodes() vs .node(), the singular version of which is used on the following line.)