fix(core): optional encoding hash in feature tests
The feature tests now need to use optional chaining whenever they look at the encoding hash. This is all in order to handle the specific case of a text mark that's specified using mark.text instead of encoding.text – a bit unfortunate, but it's not worth worrying about a cleaner syntax option right now.