chore(core): reorganize metadata function arguments
The rest of the library tends to pass the specification as the first argument, so changing this is more internally consistent.

This also reveals a hidden mistake – the outer level raw data argument was going unused by the transplant functions. This was not flagged by the linter because the specification is used and appeared later in the argument list, so the linter needed to tolerate the intermediate argument.

This makes sense, because the raw values are already available from the specification, either from specification.data.values or retrieved using the equivalent values() helper function. Passing it as a separate argument doesn't accomplish anything.