chore(core): check for specific field matches
It's not enough to check whether all fields match, as with the earlier version of the function which returned a boolean. You need to know which fields match in order to make an informed decision about whether to copy the value as metadata, so instead it needs to return an array of field names.