chore(core): field inputs for pick function
The pick function previously assumed a set of static fields which would be copied between the source and target objects, but that's not granular enough. In cases where some of the fields don't match, they need to be excluded, which means the pick function needs another argument telling you which fields to copy.