chore(core): lookup function
Data that has been aggregated by layout generators or transformed for internal use is frequently subtly different in structure. Most notably, property names at input may have been renamed with static keys for internal use, such as turning specification.encoding.color.field into the property "key" for a pie chart layout. This works well for rendering, but in order to transfer metadata between two objects, we need to be able to compare the objects, and that's harder if the property names have changed. The lookup() function goes the other way – it takes a version of a data point from later in the life cycle and converts it back into something approaching the original input so they can be more readily compared.

Most likely it will need separate lookup object creation functions for different types of charts, since they are structured differently – a pie chart doesn't have a nested series but a stacked bar chart does, and so on.