chore(core): add core encoding channels function
Not every encoding channel is worth comparing. The quantitative value can be ignored, for example, and some of the channels are the very metadata this module is trying to juggle. The coreEncodingValues() function is intended to be a reusable way to determine which encoding channels are relevant to the problem of transferring metadata fields.