chore(core): rename references to layout generators in metadata handling
Metadata handling is necessary not because these charts use d3 layout generators (which is also true) but rather because the spec requires that they aggregate incoming data. If there are two objects at specification.data.values which both have the same color encoding value, they'll be added together, and the total is used for the size of the pie slice. The metadata function is about resolving any possible incompatibilities between the input items and the aggregate total item; it doesn't actually rely on the layout generators at all, as you can see from the fact that the tests operate on the output of the data() function.