chore(core): optimize tooltip content function
Move functions to the outer scope of the factory and memoize the return value. It isn't necessary to recompute these things for each datum – just once per specification is enough, and then you can reuse that single content function on every subsequent datum. You can memoize both the creation of the content function by the factory and then also the content function itself.