chore(core): don't use typeof in undefined checks
Removes typeof checks everywhere a conditional checks for an undefined value. This was once a historical best practice because it used to be possible to change the value of undefined, but that's no longer the case in modern browsers. This change also means that an undefined check with a variable that has not been declared will throw a ReferenceError. That used to be sort of scary, but it's more disciplined to allow them – those errors should be early and catastrophic so as to encourage code styles where they are still impossible because they are appropriately guarded against.

In theory, this does make it a bit harder to get inside the conditionals where errors are thrown, which is a bit of a shame – instead of the helpful error message text, you'd get a cryptic ReferenceError, because the ReferenceError blocks the code path into the better error message!

But none of this matters, because the linter should prevent this kind of mistake anyway.