chore(core): change aria-label tests
The Vega Lite specification calls for the aria-label attribute to match the tooltip content by default. With extent descriptions added, some of the aria-label attributes no longer match exactly, because they also have the extra string appended. This is arguably a partial violation of the Vega Lite specification. Data points that are not the minimum or maximum values remain fully compliant.