chore(core): generic description for layered charts
When you layer a line over a bar chart, is it a line chart or a bar chart? This can't be reduced down to simple language, so in cases where charts are layered it's better to default to generic language like simply saying "chart" instead of trying to specify a chart type.