fix(core): disabling tooltips shouldn't disable descriptions
The tooltip content function is used to populate the text of the aria-label attribute for each mark, but the description function erroneously checks the tooltip settings in the marks before retrieving the tooltip content. These should be separate; that is, even if tooltips are disabled, the mark can use the tooltip content function to look up content for the aria-label value.