chore(core): reuse encoder functions inside bar mark renderer
Lift a few functions up to the higher level scope of the bar mark factory so they aren't needlessly recomputed for every mark. This should already be memoized, but memoization does have a little bit of overhead in computing the lookup key, and we can avoid the problem entirely by simply using scope.