chore(core): faster overlap test
The test for overlapping nodes was overly thorough at the cost of performance – it would compute the cartesian product of all input nodes, checking every possible pair for overlap with quadratic performance. This is overkill because it's not possible for the third node to overlap with the fifth node without also overlapping the fourth along the way – it's only necessary to check adjacent nodes. By assuming the input NodeList is sorted by position, it's possible to shave a few milliseconds off the overlap test function alone.