chore(core): clear chart content before rendering
Clear out all old chart content from the DOM before running the initializer. It looked like this was already happening in init.js, but once you add the detached node helper, it's just clearing the detached node, which is already empty. It's a bit duplicative to also do it in the chart wrapper, but probably fine.

Without this fix, the same chart content will render multiple times into the chart node. This bug didn't show up in simpler situations where the chart is simply rendered, but it's visible if you try to poll a data source and then rerender in place in order to update.