fix(core): correct tick alignment for horizontal time series bar charts
Most of the time it's preferable to use the data-level helpers instead of specifying channels because it gets you layout flexibility, but that logic doesn't hold inside the axis renderers, which need to specify the channels they render. Making this overly generic resulted in the axis being shifted around by the width of the bar even in cases where it was actually the other axis that was temporal and thus the shift was not necessary. This artifact was most clearly visible in a horizontal time series bar chart with x as the quantitative channel.