chore(core): prefer HCL color space
HCL is similar to HSL but is perceptually uniform, adapted for the human eye, which makes it a better choice for data encodings. Support for lab() colors has nominally been added to CSS, but in practice support for that is not yet widespread so for now it's better to compute the color using a D3 method.