chore(core): remove link events and custom handlers
This legacy behavior was always overengineered. Just open the link! No need to emit an event for it and then have the interaction module try to figure out whether a special handler has been enabled for that event. (This also cleans up config.js, which was the wrong home for this option that should not have existed at all – it should have been supplied by the consumer.)