chore(core): remove node overlap test function
Now that the axes don't try to avoid text overlap in the ticks, this function is unused, which is for the best – it's one of the slowest things to render. Maybe it will be worth reintroducing someday if another version of that problem emerges, but for now we no longer need it.