feat(core): configurable point fill
Style point stroke and fill.

These need to be handled separately for points overlaid on top of line charts. Standard Vega Lite overlays point marks with a layer, but in multiline charts that can lead to a strange occlusion artifact whereby the point marks clip the paths of lines other than the line they correspond with. To avoid that artifact, the point mark function does double duty. One mode of operation runs for line charts and applies the styling to the parent series node, and the other mode of operation styles the point marks directly.