chore(core): limit brightness changes for tritanopia
The accessibility focused color palettes are closer to the brightness extremes, likely as a result of limited hue range being available. Because of this, running luminance adjustments with the default k value of 1 is too severe and leads to some colors turning into pure white. Making the k value adjustable and then providing different values for each variant solves this. (Unfortunately the third argument means it can't just be a callback for the array function even when the k value is the default, since .map() will pass the original array as the third argument.)