chore(core): add error handling to detach helper
It helps to have a layer of error handling in the detached node helper because some things that require live DOM nodes will not work properly once they are detached – positions computed using the .getBBox() method, for example. The name of the failing function is reported as part of the error message when possible, but for anonymous functions that will be an empty string.