chore(core): filter style helper
It's likely that this pattern will soon be reused for other styling tasks, so it will probably be helpful to factor out a helper function which can quickly identify a subset of rules based on a filter predicate.

For now this is missing the type annotation for the filter function input because TypeScript doesn't seem to handle that well.