fix(core): improve tick selections
Styling needs to affect all ticks, which calls for .selectAll() instead of .select() here.

In addition, it's more appropriate to run the selection outside the tick manipulation function so it only needs to be done once and downstream functions can safely assume the incoming selection consists of the content described by the function name.