chore(core): more specific bail out conditions
As with other style rendering functions, null and undefined have slightly different meanings in the API so it's useful to mention them specifically in the source code. An undefined value means there are no styles to set; a null value means there is no node to which to apply the styles.