feat(core): apply transforms to all data functions
Despite the new wrapper layer in the data() function, it's actually not possible to run the transforms there. This is because they operate at the raw data layer, whereas by the end of the data() function the specification data has been converted by layout generators. Since those layout generators use static keys, it's actually impossible to map them back for the purposes of a transform defined in terms of the source data.

The alternative is to insert transformValues() into all uses of the values() function. This is injecting a lot of complexity into what was previously a simple lookup, but memoizing the values() helper and bypassing transformValues() entirely with the identity function when there are no transform filters to apply should both help keep things fast and lightweight.