fix(core): filter against calculate transforms without mutating data points
The previous implementation of this mapped over the input data with transformDatum() before running the filter. This has the unfortunate effect of actually mutating the data points since the calculate transforms are spread onto a new object when the calculate transform function returns. And that matters, since throughout the library memoization is applied by reference, and datum references are used to look up series values from the category helper.

To get similar functionality without mutation, we can change the functions instead of changing the data. Instead of mapping over the data points, map over the predicate function, and replace each with a wrapper function that runs the original predicate and then also runs the original predicate a second time while examining an altered version of the datum object as created by the transformDatum() function.

This is logically comparable to the way encodingValue() falls back to checking against the output of the calculate transform if no matching fields are found on the datum object.