fix(core): transform composition sequence
Previously there was no need to worry about transform composition because there was, more or less, only one type of transform available for each of the values and datum scopes; you could more or less figure out that you wanted to apply all the filters, or sample down to the smallest count. With more transform operations, however, it becomes necessary to consider the composition behavior because there may be interactions between them; a filter followed by a sample could produce a result that is quite different from the same sample followed by the same filter, applied in a different order. Vega Lite specifies that the transform array should be applied sequentially.