chore(core): pass format string directly to format functions
The format string can be stored either at channel.format or at channel.axis.format. It will be cleaner to resolve this once at the outer level, which means the format functions should take the format string as an argument directly as opposed to using the entire config object as the argument.

Using d3.format and d3.timeFormat and d3.utcFormat directly is tempting, but the wrapper functions still have their purpose, namely that they supply a default formatter (e.g. Date.prototype.toUTCString(), for example) in cases where a format string is not provided.