chore(core): handle undefined axis objects
Just like with null, axis objects that are undefined can't be memoized. In these cases the axis is not disabled, however, so the identity function is a better fallback than a noop, since there will probably still be axis ticks that require text formatting. The slight difference between the noop and the identity here is arguably illuminating?