chore(docs): explain formatting based on axis and channel
The names of these functions are a little confusing because frequently the formatting instruction at channel.axis.format is used for formatting things other than axis ticks, including primary data content which you might think would be "encoded by" the channel rather than the axis. The channel formatting is only used in a couple narrow cases. This is reflected in the type signature, but it's curious enough to deserve an explicit explanation.