chore(tests): remove extra color encoding from categorical bar fixture
A null field doesn't mean anything. This is a vestigial artifact of the days before sensible default behaviors were in place – to opt into the default colors, you needed to create a color encoding that didn't do anything in order to trigger that code path. This is no longer necessary, and it's not valid Vega Lite syntax.