chore(core): pass data object as fetch argument
A single specification can contain multiple data resources, such as in layers, and the custom fetch function needs to be able to resolve them all. It's not sufficient to pass the specification as the argument – it needs to be the individual data object, as many as are present in the specification.