add demo page as development sandbox
Adds an index.html file to the top level of the project which can be used for local development. To see it, run a local web server (try `npx http-server` or `python3 -m http.server`, for example) and then visit the local URL it provides.

Charts can be changed by changing the argument to specificationFixture(). See test-helpers.js for a list of the provided fixtures, and feel free to supply any other Vega Lite JSON as the first input to the chart() function.

This commit will cause git linting to fail. That is intentional! Please do not edit this message to make it pass continuous integration – instead, drop this commit.