chore(core): asynchronous rendering function variant
Add a wrapper function to convert a synchronous rendering function of the sort which would run with inline data into an asynchronous rendering function which first queries for remote data.

It might seem simpler to begin to treat synchronous renders as possibly-asynchronous, but that complicates the test library substantially. Instead, branching in the wrapper function allows the asynchronous functionality to be used when it's necessary while retaining the synchronous behavior as the default, and thus the tests are simpler. In other words, this will render as synchronous when it is possible to do so, and the tests will continue to be written against that behavioral option since they use inline data.