fix(core): catch extra colors in alternation
Because d3.zip() is driven by the length of the shortest array, it's possible in cases with an odd number of colors for the final color to be dropped. This manifests as colors being used redundantly, since the scale will cycle back to the beginning of the output range array. In these cases, we need to explicitly re-add the final color to the output of the function.

There's probably a more elegant way to write this, but all of them are considerably more verbose than continuing to use d3.zip() and following it with a short corrective mechanism.