chore(tests): skip testing temporal encodings
There's an off-by-one error affecting temporal encodings which is the result of gnarly shared state. Disabling these for now and will address in a follow up pull request, since it's better to have some protection in the meantime, and having some tests in place for keyboard navigation in other charts will make it easier to investigate the problem.