chore(core): handle base encodings in encodingTest() function
Functions like encodingChannelQuantitative() rely on encodingTest() to find a single channel. This could be confounded when a single dimension is expressed as multiple channels, such as y and y2 to describe an area range along a quantitative axis. In these cases it's safe to return the primary encoding channel since the secondary version shares its behavior.