chore(core): local minimum and maximum for bar width
This is a legacy configuration setting, left over from the days when sizing boundaries needed to be more explicitly enforced because edge cases might do strange things. For the most part it's no longer necessary at all, and at the very least it doesn't need to be retained as a configuration setting – the static value can be moved into the function scope.