feat(core): step calculation
The logic that has been described as "bar width" up until now can also be reused in other con texts where categorical grouping along a position dimension is useful. It's useful for it to have a more generic name for those uses, because referring to a bar will be confusing if there's no bar in sight!

On the other hand, the inverse is also true – describing something abstractly when you want the code to convey a simple concept like "bar width" is needlessly complicated.

These two things are obviously in tension, so to split the difference, an extra variable binding:

- rename the existing barWidth() function to step()
- alias the step() function back to the name barWidth()
- export both step() and barWidth()
- at the call site, use either step() or barWidth() depending on what seems more natural in context